Step towards greater encapsulation and modularity of the feature modules The Distribution.Server.Feature.* modules are the way we add functionality to the hackage server in a nice modular way.
For this to be robust, these features need to properly encapsulate
the state they manage, and provide APIs for other features to use
that state in disciplined ways. Currently the AcidState stuff is
global, rather than partitioned between features. This is primarily
for historical reasons (the original happstack-state forced the state
to be application global), and for ease of porting to the new
acid-state package, we kept the same pattern of global shared state
(in the Distribution.Server.Acid module).

Obviously we want to split this up so that each feature encapsulates
its own state. The first step to doing this more easily is to change
the features to use a "first class module" style, where all the
exported functions are part of a feature record. This will make things
much easier since as we will not need to pass state tokens around to
every query or update. e.g instead of:

foo :: UsersFeature -> ...
foo usersF ... =
  ...
  users <- queryGetUsersDb usersF

it'll be simply

foo :: UsersFeature ->
foo UsersFeature{..} ... =

  users <- queryGetUsersDb

And indeed, most functions don't need to take the feature arguments
explicitly, just the top level initBlahFeature functions, with all
other functions in the module defined in the where clause.

This may seem a little odd, but in my experiments in trying to
encapsulate the state properly, this seems like the best approach.