Refactor: reify state components
This should make it harder to forget to provide backup methods (and easier to
check that they are provided, and which are provided). It also simplifies the
HackageFeature data type.

Next step is to see if we can change the definition of StateComponent so that
we can express better what the backup should do.