Make export pure
The only reason export lived in IO was because it needed access to the blob
store. Now instead we define 

data BackupEntry =
    BackupByteString [FilePath] ByteString
  | BackupBlob [FilePath] BlobId

This has various other advantages: we now have the option of storing blob
entries differently (or indeed not at all) when we generate the tar file.
For now, the export format is still the same, but at least we have the 
option now of treating blobs differently.

Note that there was already an ExportEntry (as opposed to BackupEntry) type
which was isomorphic to the (new) BackupEntry type. I have removed that type
completely, and removed some of the conversions between them.