Fix the 'restore' process
- Not all features implemented 'restore' after the refactoring, causing 
  restore to fail
- 'importTar' assumed a single RestoreBackup per feature, while we now have
  multiple (though, thankfully, they implemented Monoid)

Note that 'testRoundtrip' is still broken, but it turns out it was already
broken even before the refactoring. The problem is that testRoundtripByQuery
returns a check which compares the state of the database using the AcidState
handle which corresponds to the *initial database* (before the export and
reimport), and hence is comparing a database to itself and always returns
success. It should instead use the handle created after the re-import, but that
handle is not yet available when 'testRoundtrip' is called, so this requires
some re-engineering of how tests are set up. Will fix this separately.