Add 'resetState' to StateComponent
resetState has type

  StateComponent -> BlobStorage -> FilePath -> IO (StateComponent st)

and basically creates a new state component with the initial state. This is now
used consistently throughout the initFooFeature functions to initialize the
state, so that many initFeature methods now simply look like

initFooFeature <some params> = do
  st <- fooStateComponent ..
  return (fooFeature <params> st)

fooStateComponent = do
  st <- openLocalStateFrom ...
  return StateComponent {
      ...
    , resetState = fooStateComponent
    }

This is a nicer separation of concerns, but more importantly, this is a
stepping stone towards fixing 'testRoundtrip'.