Remove testRoundtrip
StateComponent now provides enough informaiton to define testRoundtrip
generically for all features (just equality on states). This is possible
because the check for *blobs* will not be necessary anymore, as we will take
care of this generically, too (we don't yet, though).

This now throws an error in test-backup in the import/export for Core;
presumably this error was always there, we just didn't notice it becauset
testRoundtrip was set up incorrectly. (I think the error is not so important,
rounding error for dates.)