Overhaul all the caching Separate the previous uses of caches into two kinds: 1. simple in-memory vars and 2. proper caches. For the former we just add a simple strict wrapper around MVar, with read/write/modify operations, while for latter kind we use asynchronous caches that are constructed with their update action.
Then update the features to use the appropriate kind of cache.
Additionally, the old async cache infrastructure had a feature to
disable updating of caches for the benefit of bulk import operations.
The corresponding thing in the new infrastructure is a tunable knob
to delay cache updates so that more changes can accumulate before doing
the expensive cache update operation.