More renaming. Disentangle some dependencies
What we want to do is move the 'lookupTarIndex' to the feature API of the
package contents feature, so that we can cache the tarball indexes. However,
this will introduce a circular dependency between the package contents feaure
and the package candidates feature. 

More generally, there is quite a bit of overlap between these two features. We
should abstract out the common functionality and then disentangle the circular
dependencies. This will require a bit more thought though so not doing that
right now.