Remove dependency of PackageContents on PackageCandidates
This means there is a bit of duplication between PackageCandidates and
PackageCandidates (serveTarball and friends), but we can factor that out later.
Importantly, both PackageContents and PackageCandidates can now maintain their
own state (cached TarIndexes) so that we can start to get rid of the stuff in
Util/ChangeLog (which reconstructs the TarIndex on every request).