Add packageTarball and packageChangeLog to PackageContents
At the moment these are simply defined in terms of
lookupTarballAndConstructTarIndex and lookupChangeLog, but those functions are
now no longer exported (RecentPackages now depends on the PackageContents
feature to get a handle on a package changelog). This means that we are now
finally in a position that we can actually start caching these tar indices.