Factor out common PackageContents/PackageCandidates code
There is still a bit of duplication, but this is "uninteresting" code (in the
sense that there are no real design decisions involved, just getting
information from various places). In particular, the code that finds a
changelog in a tarball (looking for ChangeLog, CHANGELOG, etc.) is now in a
separate module.