Have the Documentation use the TarIndexCache
This means that we don't keep all the tar indices in memory but deserializable
them on every request, thereby saving memory (and the Documentation feature is
simpler now that the tar index caching is done in a different feature).

I noticed however that the build client no longer works; in particular,
uploading documentation does not work (I think the URL is wrong;
.../package/doc rather than .../package/docs, and it tries to upload a .tar.gz
while we expect a .tar).