Core cleanup (1/N)
Core had

  withPackageId   ::              DynamicPath -> (PackageId   -> ServerPartE a)   -> ServerPartE a
  withPackageName :: MonadIO m => DynamicPath -> (PackageName -> ServerPartT m a) -> ServerPartT m a

But both of these are an instance of a much more general function

  withPackageInPath :: (MonadPlus m, FromReqURI a) => DynamicPath -> (a -> m b) -> m b

So I have removed the first two in favour of the withPackageInPath. Moreover,
withPackageInPath belongs in CoreResource more than in CoreFeature -- it
doesn't access the package database, but simply extracts a bit from a URL.

CoreFeature still contains a whole plethora of withXXX functions; the logical
place for these is indeed in Feature rather than Resource, but we don't need
quite all the different combinations; we can offer more general functions. Will
do that next.

The immediate reason for all this cleanup is that we want the PackageCandidates
feature to ultimately implement the core API so that other packages (build
reports, documentation, tags, mirror, versions, ..?) can work with either.