Add guardValidPackageXXX
In our relentness quest to bring Core and Candidates closer together, add
guardValidPackageId and guardValidPackageName to CoreResource. Many features
just need to check if a package ID is valid, rather than get access to the
actual package (BuildReports being one example). This means that these features
can easily work with either Core or Candidates (which, after all, use different
types of packages).

This also addresses the issue mentioned in the previous log, where we used the
idiom

  _ <- lookupPackageXXX

to check that a package exists.