Overhaul the user account and authentication code Remove the htpasswd feature, simplifying the UserAuth stuff so that there is now just one way to store password hashes. The legacy htpasswd stuff will be reinstated in a separate legacy password feature module.
Simplify the representation of user status, we now have just enabled,
disabled (with and without auth details) and deleted. This leads to
simplifications in the Users db representation and operations. Also
remove the silly requireName thing which automatically created
historical accounts on demand, instead mirror clients will be expected
to create user accounts if they want to preserve author info.