Redo the maintainer links and page based on feedback The feedback was that the links for maintainers were confusing for ordinary users because it is not clear if they are allowed to use them or not, and "maintain" is not very descriptive as a link name.
This patch removes the tags edit link from the package page. It moves
the maintainer edit link to the bottom of the page and make it clearer
that it is only for maintainers and trustees. The new maintainer page
is now a lot clearer (and uses a template so easier to edit).