Reduce memory footprint of DownloadCounts (1/n)
Change the DownloadCounts feature so that it only keeps today's statistics in
memory, and everything else on disk (and simplify the API at the same time).

This isn't quite done yet though, because backup/restore has not yet been
implemented, and will require some changes to the backup/restore infrastructure
because we have no way of backuping the on-disk statistics at the moment.