Make lexer monadic (avoid use of 'error')
Importing package metadata was failing because the packageAddHook added by the
Search feature was parsing the package description, and if this had a lexical
error (as far as the Haddock lexer was concerned) it would throw a runtime
error which would eventually result in a Internal Error 500 error. The lexical
now returns a value in the Either String monad, just like the parser, so that
we can run 

    Haddock.tokenise >=> Haddock.parseHaddockParagraphs

to savely lex and parse, without worrying about exceptions in pure code.