Audit for UTF8 safety (and remove json)
This is a longish commit which removes a lot of uses of Data.ByteString.Char8
and Data.ByteString.Lazy.Char8 in favour of UTF8 safe functions. There are a
few (hopefully legiminate) uses of D.B.*.Char8 left, most notably in
Client.ParseApacheLogs and Client.UserAddressesDb, and a few small uses in
other modules, marked with why they are ok. A handful of modules are not yet
checked or changed; they have been marked as TODO. 

Also removed a bunch of compiler warnings in the process.