Build docs for most recent packages first
This is a slightly subtle change to the doc builder: before the patch, we would
find all packages that needed their docs build and then order them so that we
would build the docs for the most recent packages first. However, this meant
that if package A already had docs for its most recent version and package B
did not, we would build the docs for A's second-most-recent version before
building B's most recent version. Now, we *first* sort the list of packages so
that we consider the most recent version of each package first and only THEN
filter them by those versions that need their docs built.