Use new http session for every package doc upload
Before this patch we would use a single http session for the entire doc builder
execution. However, since doc building can take quite a long time this would
cause long periods of inactivity, resulting in a 
  
  <socket: 3> resource vanished

error. To avoid this, we now use a http session every time we upload: so, we
build docs, use a new http session to upload the docs, repeat. This seems to be
a lot more reliable.