Improve statistics reported by doc builder
Previous the doc builter reported which packages had their most recent package
built (yes/no) and which packages had all versions built (yes/no). However,
"did this package version" is not a yes/no question, it is a "yes, it built",
"yes, we built, but we failed", or "no, we didn't try yet" question. To monitor
progress of the doc builder it is therefore more useful to take this ternary
answer into account.