Only retry cabal test for the moment in travis config
Travis seems to have random network problems and their current recommended workaround is to retry a few times. I think we only want to do this for the tests (which is where we've seen the problems) and not for the main code compile check (where we'd end up getting the same build failure three times or whatever).