Make README useful
I'd really like to see the README contain useful information on how to
quickly get started with the hackage server. In this commit I started
with some of the info on the wiki page, and fixed/updated this info in
several cases. Some of the updates I made:

* Add travis-ci build status to README
* Update commands to use `cabal`
* Fix steps for mirroring since wiki seems to use outdated instructions

If we merge this, I'll also delete the outdated info on the
wiki. Thanks - let me know if you think anything else should be done
to merge this or if you disagree with the changes I made!