Move a couple resource handlers from html to user feature
Previously the user feature didn't serve the get/put/delete methods
directly but exported methods that the html feature used to implement
them with html responses. Move it around so it's sane: methods using
JSON formats in the user feature and no wrappers in the html feature.
Use in html pages relies on the form wrappers stuff.