Split feature initialisation into an additional phase
The first phase of each feature initialisation is now dependent only
on the server env and not on the other features that the feature
depends on. The dependent features are supplied in the second phase.
This makes running them for test/debug purposes. The first phase is
typically the most expensive as that is when we load data from disk.
So the first phase being expensive and independent of all other
features opens the door to further parallelising the server startup.