Require new bytestring with perf bug fix for server startup
Loading the acid-state disk snapshots was running slowly due to the
crc16 impl in acid-state using lazy bytestring foldl' which -- it
turned out -- was implemented incorrectly (which as one of the authors
of the bytestring code is rather embarassing). So with the 8-year old
bytestring bug squashed, we just need to require the new version to get
a free speedup on loading large acid state snapshots.

This eliminates a stack use spike during server startup (one of the
larger data stores).