Fix a mistaken migration that reversed all .cabal file revisions
A recent patch changed the representation of the package info, including
the sequence of revisions to the .cabal file. The intention of this
change was to simplify the representation and make it less confusing.
Sadly the confusion bit me in the migration, and I accidentally reversed
all the .cabal file revisions.

This patch adds a follow-up migration that puts them all back in order.

Fixes issue #305.