Render readme and changelog as html, but not inline for now
Disable rendering the readme inline next to the package description.
It works, but unfortunately for many packages the result is confusing
because they have both a description and readme that mostly duplicate
each other (or they have an old readme that the authors have forgotten
about).

Hopefully we can improve the heuristic of when to show the readme and
then re-enable. Also, readmes tend to be very long so we should
implement a "top N lines" system, and link through to the full readme.
Otherwise the long readme pushes the rest of the info off the screen.

However, we do now link to the readme if it's present, and render it as
markdown. We also do the same for the changelog.