upload: provide plain text response API
This adds an API to upload a package and get a plain text response with errors or
warnings. Previously, this was only available through a legacy url for packages and
not available at all for candidates.

The plain text API is useful for command line tools, such as cabal-install, for example.