Refactor package page generation to use templates.
Everything about the package page should be functionally identical to
how it always has been, but offloads most of the styling to a template
instead.

There are also now a few conditional variables provided to see if
particular bits of information is available for a given package. This
allows more of the conditional logic to take place in the template as
well, hopefully making it easier to tweak the UI's design and
responsiveness without necessarily needing to dive into the source.

Changes also include a few efforts to reduce the line count in Html.hs
(which, as a comment notes, is pushing 1k lines). Most of the functions
from Pages/Package.hs are still used, but a few of them needed slight
changes, and are duplicated in the new PackageFromTemplate module to
avoid any regressions (e.g. on the maintain-candidate page).

Signed-off-by: Duncan Coutts <duncan@community.haskell.org>