Introduce 'TUFInput' into the security state
This will enable us to reproduce precisely the same TUF files after a server
restart or a backup import, which is essential for the backup roundtrip. This
commit _just_ introduces the new state; I've committed this separately from
actually making using of this state because this bit requires a (simple)
migration.