Simplify TUFFile and Introduce SafeCopy instance
The SafeCopy instance will be necessary to return a TUFFile from an acid-state
transaction; the simplification is that we removed the expiry time. The latter
was originally intended for cache control, but we don't want to use expiry
times for that, so it's now redundant. Removing it allows for some
simplifications in the code, removes a possible runtime exception (not all
files _have_ an expiry time), and if we now introduce a SafeCopy instances it
makes sense to do it now or else we might have to introduce migrations later.

This is a bit of an annoying change, because `TUFFile` used to have a phantom
type argument indicating what kind of TUFFile it was (this type argument was
instantiated with the corresponding type from the hackage-security library).
Sadly, `deriveSafeCopy` is not very smart and when applied to `TUFFile` it
added a spurious `SafeCopy` requirement on the phantom type argument. To avoid
this problem, in this commit we remove the phantom type argument and instead
introduce a bunch of newtype wrappers.