Update to latest hackage-security
The hackage-security library now uses Int54 instead of Int everywhere (for
annoying JSON reasons). In this commit I prove a SafeCopy instance for Int54
which translates to Int, so that we remain backwards binary compatible. I do
change to Int54 in a few datatypes, but only in datatypes that are not (yet)
serialized on the running server.