Remove cabal-version:>=2.0 entries from legacy 00-index
The new cabal-version:2.0 format introduces new syntax
but unfortunately confuses old cabal-install versions.

On the bright side, versions prior to cabal-1.24 only support
the 00-index legacy index. While cabal-1.24 uses the 00-index
by default (users have to opt into secure repositores in order
to access the new 01-index). Since old cabal versions
can't deal with cabal-version:2.0 entries anyway, there's no
harm in hiding them in the legacy index.

So with this server-side workaround we should be able to
mitigate the issue for the majority of the legacy-user base.

The implementation of this workaround is not very principled,
but due to the urgency, it'll have to be refactored some time
later...