Make it responsive
Right now, hackage is not responsive, which makes UX terrible.
For example, when viewing the page from a small device, the header
links overlap and the user no longer sees them all. Also, it is
annoying that the #content takes the same space in a small screen
as in a big one, as it results in very small sentences.