Move title (synopsis+description) out of the navigation header
This was suggested in
https://github.com/haskell/hackage-server/pull/648#issuecomment-354275785 :

Paraphrasing from the comment: Having the title in the header is a bad
idea as its size is always different, which makes the header size
unpredictable and ugly often enough. Thus we move the title to the
content area (which also makes the title feel more associated with the
content), and let the navigation header always show the
`Hackage :: [Package]` caption across all of Hackage.