Change color of :visited links
The previous color, as pointed out by jirassimok, showed
more contrast than the normal links color, which is not how
it should be. This new color was suggested by Adobe Color Wheel.
It is more faded, less contrasting and less aggressive on the eye as
well,
making the page still look good when all links are visited.