Add new simpler and friendlier favicon
The current favicon doesn't work well in such a small space. Besides
that, its transparent background doesn't work well in browsers with
dark themes. We agreed to change to a favicon with a haskell-purple
background color with the haskell logo. This commit also changes how
the favicon is exposed to the browser, by using <link> pointing to
a .png file, instead of *.ico as the first appears to be much more
size/quality efficient. Since the favicon declaration becomes explicit
there is no need to serve the it as a root-level file.