Close gap in case-clash validation
This is a follow-up to 83e7c4d8d2a27275fd928b37c7e603a91b4f41b3

While at it, turn package names into hyperlinks in error message.

NB: This is still not ideal; but it should enforce the invariant we
need for the primary package index at least.