Fix handling of curl's optimised digest handshake
Commit f8e0060aa150a50852e1efd9e20faa9a2d8d4125 deferred the authentication
check but this didn't interact well w/ the way the anticipated digest auth
handshake is implemented in `curl` (which sends a stripped down version of
the actual request to initiate the digest auth handshake).

So this commits adds back the eager authentication check, but merely
requiring *any* known user account.

Fixes #772