Vendor ListT, removed in transformers-0.6 (Fix #1010) (#1027)
`ListT` is used in a utility module for package search.  The functions in the utility module implement some heuristics to extract words from package identifiers.  The module is undocumented.
It does not seem worthwhile at this point to reverse engineer this module and try to implement it w/o `ListT`.
Rather, I embed the `ListT` monad transformer code here so we do not depend on getting it from `transformers` (where it has disappeared from since `>= 0.6`).
This also silences the deprecation warnings.