Minimal migration to Cabal-3.6.3.0
This patch makes the minimal changes to compile with Cabal-3.6.3.0.

Cabal-3.6 intends a semantics of the caret operator ^>= that is more
subtle than expanding it into its version interval.
We do not pick up this intention in this patch, but fall back to

  Distribution.Types.VersionInterval.Legacy

See https://github.com/haskell/cabal/issues/7916 for a in-depth
discussion.