accept additional quick-jump hashes
Accept all known-good `quick-jump.min.js` and `quick-jump.css`
hashes.  This function also partitions the check by path: only the
js or css hash list is checked (whichever is relevant).

A modest refactor makes the code a bit easier to follow and makes it
easy to add hash checks for oher paths in the future, if needed.